pyproject.toml
[tool.poetry]
name = "SID-api"
version = "0.1.0"
description = "API and database management for SID project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["standard"], version = "0.115.0"}
uvicorn = "0.30.6"
sqlalchemy = "2.0.35"
alembic = "1.13.3"
psycopg2-binary = "2.9.9"
requests = "2.32.3"
asyncpg = "0.29.0"
asyncio = "3.4.3"


[tool.poetry.group.dev.dependencies]
pytest = "8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
